Split PaymentModule CampPresenter
