+++ app/AccountancyModule/PaymentModule/templates/CreateCampGroup/default.latte
@@ -1,3 +1,5 @@
+{varType Model\Event\Camp $camp}
+
 {block content}
 <h1 n:inner-block="title">Nová skupina plateb <small>{$camp->displayName}</small></h1>
-{control newGroupForm}
+{control form}
