+++ app/AccountancyModule/PaymentModule/templates/GroupList/default.latte
@@ -31,7 +31,7 @@
 
                         <div class="dropdown-divider"></div>
 
-                        <a n:href="Camp:default" class="dropdown-item">
+                        <a n:href="SelectCampForGroup:" class="dropdown-item">
                             <i class="fas fa-campground"></i> Tábor
                             <span class="small d-block">Umožnuje načíst účastnické poplatky.</span>
                         </a>
